THING THREE
There's a bunch of clients marked as "Third-Party Reimplementations" that is in fact a custom discord launcher with a modded version of official discord client. And this is a problem, because when you scroll through "3rd-p impls" you expect to see repos that are actually reimplementations -- like, with custom UI and all that stuff.